footer layout changes and profile
